+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Include token in subgraph even if not covered by segmentation node #334.
+
 ## [4.1.1] - 2026-01-13
 
 ### Added